refactor(react): restructured navigation file system
